// src/app/api/user/models/user-request-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserRequestDto {
  addressCity?: string;
  addressCountry?: string;
  addressStreet?: string;
  addressZipCode?: string;
  countryCode: string;
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  phoneNumber: string;
  roles?: Array<'USER' | 'ADMIN' | 'SELLER' | 'SUPPORT'>;
}